week6/1-Market-REST/market/website/serializers.py
from rest_framework import serializers
from .models import Offer


class OfferSerializer(serializers.ModelSerializer):
    class Meta:
        model = Offer
        fields = ('id', 'title', 'price', 'description', 'created_date', 'status')
