week6/1-Market-REST/requirements.txt
decorator==4.0.11
Django==1.10.5
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
djangorestframework==3.6.2
factory-boy==2.8.1
Faker==0.7.10
gunicorn==19.6.0
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.1.0
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
psycopg2==2.7
ptyprocess==0.5.1
Pygments==2.2.0
python-dateutil==2.6.0
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
wcwidth==0.1.7
